requirements.txt
cython
tqdm
yacs
numba
numpy
filterpy
chumpy
Pillow
trimesh
pyrender
matplotlib
json_tricks
scikit-image
opencv-python
scikit-image
scipy==1.4.1
tensorboardx
smplx==0.1.28
torchgeometry
git+https://github.com/philferriere/cocoapi.git#egg=pycocotools&subdirectory=PythonAPI
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git
ultralytics==8.0.24
